+++ fairifier/__init__.py
@@ -9,4 +9,4 @@ try:
 except ImportError:
     pass
 
-__version__ = "2.0.3"
+__version__ = "2.0.4"